separate doc reader
