+++ examples/gaia_agent/environment.py
@@ -3,7 +3,7 @@ import logging
 from tapeagents.environment import ToolCollectionEnvironment
 from tapeagents.tools.browser import Browser
 from tapeagents.tools.code_executor import CodeExecutor
-from tapeagents.tools.document_converters import DocumentReader
+from tapeagents.tools.document_reader import DocumentReader
 from tapeagents.tools.media_reader import VideoReader
 from tapeagents.tools.search import Search
 from tapeagents.tools.simple_browser import SimpleBrowser
